Renamed the `DTO` objects, and added a small comment in the `DbContext` class.
